Update phased prow logic
